refactor: replace CorrelatedTUnitLogger with internal SynchronousTUnitLogger
ASP.NET Core's built-in ConsoleLogger writes from a background queue
thread that doesn't inherit the AsyncLocal set by MakeCurrent(). We
still need a logger that writes to Console synchronously on the request
thread so TUnit's interceptor routes output to the correct test.

Replaced the public CorrelatedTUnitLogger/Provider with a minimal
internal SynchronousTUnitLogger — no dedup check, no caching, no
IHttpContextAccessor dependency. Just writes to Console if
TestContext.Current is set.
