+++ TUnit.AspNetCore.Core/Logging/CorrelatedTUnitLoggingExtensions.cs
@@ -6,26 +6,23 @@ using Microsoft.Extensions.Logging;
 namespace TUnit.AspNetCore.Logging;
 
 /// <summary>
-/// Extension methods for adding correlated TUnit logging to a shared web application.
+/// Extension methods for adding TUnit test context middleware to a shared web application.
 /// </summary>
 public static class CorrelatedTUnitLoggingExtensions
 {
     /// <summary>
-    /// Adds correlated TUnit logging to the service collection.
-    /// This registers the <see cref="TUnitTestContextMiddleware"/> via an <see cref="IStartupFilter"/>
-    /// and a <see cref="CorrelatedTUnitLoggerProvider"/> that resolves the test context per log call
-    /// via <see cref="TUnit.Core.TestContext.MakeCurrent"/>.
-    /// Use with <see cref="TUnitTestIdHandler"/> on the client side to propagate test context.
+    /// Adds the <see cref="TUnitTestContextMiddleware"/> to the pipeline via an <see cref="IStartupFilter"/>
+    /// and a synchronous logger provider that writes <c>ILogger</c> output to <see cref="System.Console"/>
+    /// on the calling thread so TUnit's console interceptor can route it to the correct test.
+    /// Use with <see cref="TUnitTestIdHandler"/> on the client side to propagate the test context ID.
     /// </summary>
     /// <param name="services">The service collection.</param>
-    /// <param name="minLogLevel">The minimum log level to capture. Defaults to Information.</param>
     /// <returns>The service collection for chaining.</returns>
     public static IServiceCollection AddCorrelatedTUnitLogging(
-        this IServiceCollection services,
-        LogLevel minLogLevel = LogLevel.Information)
+        this IServiceCollection services)
     {
         services.AddSingleton<IStartupFilter>(new TUnitTestContextStartupFilter());
-        services.AddSingleton<ILoggerProvider>(new CorrelatedTUnitLoggerProvider(minLogLevel));
+        services.AddSingleton<ILoggerProvider>(new SynchronousTUnitLoggerProvider());
 
         return services;
     }
